Event Board
Fix comments
