+++ package.json
@@ -40,7 +40,7 @@
         "@capacitor/keyboard": "^8.0.3",
         "@capacitor/preferences": "^8.0.1",
         "@capacitor/status-bar": "^8.0.2",
-        "@capawesome/capacitor-network": "0.1.2",
+        "@capawesome/capacitor-network": "0.1.3",
         "@capgo/background-geolocation": "^8.0.39",
         "@capgo/capacitor-compass": "^8.1.17",
         "@capgo/capacitor-updater": "^8.50.1",
